+++ src/pages/Color.vue
