+++ src/pages/Hash.vue
