+++ src/pages/encode/Base64.vue
