--- src/src/main.css
@@ -1,2 +1,0 @@
-@tailwind base;
-@tailwind utilities;
