--- src/src/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "tools",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build"
-  },
-  "dependencies": {
-    "2fa": "^0.1.2",
-    "@vue/shared": "^3.2.0-beta.5",
-    "clipboard-polyfill": "^3.0.3",
-    "crypto-js": "^4.1.1",
-    "date-fns": "^2.23.0",
-    "dexie": "^3.0.3",
-    "easyqrcodejs": "^4.4.3",
-    "element-plus": "^1.0.2-beta.65",
-    "i": "^0.3.6",
-    "jsoneditor": "^9.5.3",
-    "lodash": "^4.17.21",
-    "nprogress": "^0.2.0",
-    "otpauth": "^7.0.4",
-    "quasar": "^2.0.3",
-    "vue": "^3.1.5",
-    "vue-i18n": "^9.1.7",
-    "vue-router": "^4.0.10",
-    "yaml": "^2.0.0-7"
-  },
-  "devDependencies": {
-    "@intlify/vite-plugin-vue-i18n": "^1.0.0-beta.17",
-    "@types/crypto-js": "^4.0.2",
-    "@types/lodash": "^4.14.171",
-    "@vitejs/plugin-vue": "^1.3.0",
-    "@vue/compiler-sfc": "^3.1.5",
-    "autoprefixer": "^10.3.1",
-    "postcss": "^8.3.6",
-    "tailwindcss": "^2.2.7",
-    "vite": "^2.4.4"
-  }
-}
