--- src/src/views/Index.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>
-    <q-btn color="white" text-color="black" label="AA" />
-  </div>
-</template>
-<script setup>
-</script>
